pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "gitmortem"
version = "0.1.0"
description = "Instant incident postmortems from a git commit hash"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "gitmortem contributors" },
]
keywords = [
  "git",
  "postmortem",
  "incident-management",
  "devops",
  "sre",
  "llm",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Version Control :: Git",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "anthropic>=0.40.0",
  "gitpython>=3.1.43",
  "markdown>=3.6",
  "openai>=1.40.0",
  "rich>=13.7.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[project.scripts]
gitmortem = "gitmortem.cli:run"

[project.urls]
Homepage = "https://github.com/yourusername/gitmortem"
Repository = "https://github.com/yourusername/gitmortem"
Issues = "https://github.com/yourusername/gitmortem/issues"

[tool.hatch.build.targets.wheel]
packages = ["gitmortem"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["B008"]
